feat: Moves the postCreate route to another folder
